[procmgr] Move test_uuid and make_config to test_helpers module
Centralize shared test utilities so all test modules (process, shutdown,
grpc/service) import from one place instead of cross-referencing
process::tests.
